backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]
requires-python = ">=3.11,<4.0"

dependencies = [
    "openai>=1.0.0",
    "requests>=2.32.0",
    "langgraph>=0.3.31",
    "python-dotenv>=1.0.0",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# ─────────────────────────────────────────────
# ✨ 추가: src/agent 패키지를 설치 대상에 포함
[tool.poetry]
packages = [
  { include = "agent", from = "src" }
]
# ─────────────────────────────────────────────
